Refactor `ErrorPrinter` into a separate file
